+++ app/routes/index.js
@@ -4,10 +4,23 @@ const calculationRoutes = require("./calculation.routes");
 const ruleRoutes = require("./rule.routes");
 const accountRoutes = require("./account.routes");
 const faqRoutes = require("./faq.routes");
+const playgroundRoutes = require("./playground.routes");
 
 module.exports = router;
 
+router.use("/account", accountRoutes);
 router.use("/calculations", calculationRoutes);
 router.use("/rules", ruleRoutes);
-router.use("/account", accountRoutes);
 router.use("/faq", faqRoutes);
+
+router.use("/playground", playgroundRoutes);
+
+// router.use(authChecker);
+
+function authChecker(req, res, next) {
+  if (req.session.auth || req.path === "/auth") {
+    next();
+  } else {
+    res.redirect("/auth");
+  }
+}
